+++ raptor/util/linalg/repartition.cpp
@@ -496,7 +496,7 @@ ParCSRMatrix* repartition_matrix(ParCSRMatrix* A, int* partition, std::vector<in
             {
                 proc = recv_status.MPI_SOURCE;
                 MPI_Get_count(&recv_status, MPI_INT, &count);
-                if (count > recv_row_buffer.size())
+                if (count > (int) recv_row_buffer.size())
                 {
                     recv_row_buffer.resize(count);
                 }
@@ -523,7 +523,7 @@ ParCSRMatrix* repartition_matrix(ParCSRMatrix* A, int* partition, std::vector<in
         {
             proc = recv_status.MPI_SOURCE;
             MPI_Get_count(&recv_status, MPI_INT, &count);
-            if (count > recv_row_buffer.size())
+            if (count > (int) recv_row_buffer.size())
             {
                 recv_row_buffer.resize(count);
             }
